# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EpiCare AI",
    "description": "\n**EpiCare AI** is a robust and comprehensive platform designed to manage and analyze epilepsy-related data, facilitating seamless interactions between patients, doctors, and caretakers.\n\n### Core Features\n- **Admin Dashboard**: Comprehensive system metrics and doctor PMDC verification tools.\n- **User Management**: Unified profiles tailored for Patients, Doctors, and Caretakers.\n- **Role-Based Authentication**: Secure access control with PMDC verification for doctors.\n- **Connection System**: Connect patients with verified medical professionals and trusted caretakers.\n- **AI-Driven EEG Analysis**: Upload EEG sessions, process signals, and receive detailed AI-generated seizure prediction reports.\n- **Personalized Dashboard**: Get real-time heuristic recommendations based on daily habits.\n- **Lifestyle & Medication Tracking**: Log sleep, stress, triggers, and medications with adherence algorithms.\n- **Emergency SOS**: Instantly dispatch multi-channel (Firebase/WhatsApp/Email) SOS alerts to configured contacts.\n\nThis API adheres strictly to REST principles, delivering standardized responses and explicit HTTP status codes to ensure a reliable developer experience.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me/patient-profile": {
      "get": {
        "tags": [
          "Patient Management"
        ],
        "summary": "Get patient profile",
        "description": "Retrieve the current authenticated user's patient profile.",
        "operationId": "get_my_patient_profile_api_v1_users_me_patient_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "404": {
            "description": "Patient profile not found for this user"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Patient Management"
        ],
        "summary": "Update patient profile",
        "description": "Create or completely update the current user's patient profile.",
        "operationId": "update_my_patient_profile_api_v1_users_me_patient_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Patient Management"
        ],
        "summary": "Delete patient profile",
        "description": "Delete the current user's patient profile. The underlying user account remains active.",
        "operationId": "delete_my_patient_profile_api_v1_users_me_patient_profile_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "404": {
            "description": "Patient profile not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/doctor-profile": {
      "get": {
        "tags": [
          "Doctor Management"
        ],
        "summary": "Get doctor profile",
        "description": "Retrieve the current authenticated user's doctor profile, including PMDC verification status.",
        "operationId": "get_my_doctor_profile_api_v1_users_me_doctor_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorProfileOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "404": {
            "description": "Doctor profile not found for this user"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Doctor Management"
        ],
        "summary": "Update doctor profile",
        "description": "Update the current user's doctor profile details.",
        "operationId": "update_my_doctor_profile_api_v1_users_me_doctor_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorProfileOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Doctor Management"
        ],
        "summary": "Delete doctor profile",
        "description": "Delete the current user's doctor profile. The underlying user account remains active.",
        "operationId": "delete_my_doctor_profile_api_v1_users_me_doctor_profile_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "404": {
            "description": "Doctor profile not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/caretaker-profile": {
      "get": {
        "tags": [
          "Caretaker Management"
        ],
        "summary": "Get caretaker profile",
        "description": "Retrieve the current authenticated user's caretaker profile.",
        "operationId": "get_my_caretaker_profile_api_v1_users_me_caretaker_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaretakerProfileOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "404": {
            "description": "Caretaker profile not found for this user"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Caretaker Management"
        ],
        "summary": "Update caretaker profile",
        "description": "Create or completely update the current user's caretaker profile.",
        "operationId": "update_my_caretaker_profile_api_v1_users_me_caretaker_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaretakerProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaretakerProfileOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Caretaker Management"
        ],
        "summary": "Delete caretaker profile",
        "description": "Delete the current user's caretaker profile. The underlying user account remains active.",
        "operationId": "delete_my_caretaker_profile_api_v1_users_me_caretaker_profile_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token"
          },
          "404": {
            "description": "Caretaker profile not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaretakerProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "relationship_to_patient": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship To Patient"
          },
          "crisis_phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crisis Phone Number"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "CaretakerProfileOut"
      },
      "CaretakerProfileUpdate": {
        "properties": {
          "relationship_to_patient": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship To Patient"
          },
          "crisis_phone_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Crisis Phone Number"
          }
        },
        "type": "object",
        "title": "CaretakerProfileUpdate"
      },
...
      "DoctorProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "pmdc_number": {
            "type": "string",
            "title": "Pmdc Number"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "hospital_affiliation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hospital Affiliation"
          },
          "license_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Image Url"
          },
          "is_pmdc_verified": {
            "type": "boolean",
            "title": "Is Pmdc Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "pmdc_number",
          "specialty",
          "is_pmdc_verified",
          "created_at",
          "updated_at"
        ],
        "title": "DoctorProfileOut"
      },
      "DoctorProfileUpdate": {
        "properties": {
          "specialty": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "hospital_affiliation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Hospital Affiliation"
          },
          "license_image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "License Image Url"
          }
        },
        "type": "object",
        "title": "DoctorProfileUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "blood_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Type"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "primary_diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Diagnosis"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Relation"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "known_triggers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Triggers"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "date_of_birth",
          "timezone",
          "created_at",
          "updated_at"
        ],
        "title": "PatientProfileOut"
      },
      "PatientProfileUpdate": {
        "properties": {
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Male",
                  "Female",
                  "Other",
                  "Prefer not to say"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "blood_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10,
                "pattern": "^(A|B|AB|O)[+-]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Type"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "primary_diagnosis": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Diagnosis"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_relation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Relation"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "pattern": "^\\+?[1-9]\\d{1,14}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "known_triggers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Triggers"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "title": "PatientProfileUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}